+++ src/demo/src/FragmentedMPEG/isobmff/FragmentedISOBMFFBuilder.js
@@ -1,3 +1,19 @@
+/* Copyright 2020 Ethan Halsall
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <https://www.gnu.org/licenses/>
+*/
+
 import Box from "./Box";
 
 /**
@@ -5,6 +21,10 @@ import Box from "./Box";
  * wrap MPEG frames in a MP4 container for streaming MP3 compatibility in Firefox.
  */
 export default class FragmentedISOBMFFBuilder {
+  static getBoxContents(boxes) {
+    return Uint8Array.from(boxes.flatMap((box) => [...box.contents]));
+  }
+
   /**
    * @returns {Uint8Array} Generic Filetype and Movie Box information for MP3
    */
@@ -114,9 +134,5 @@ export default class FragmentedISOBMFFBuilder {
     trun.insertBytes(Box.getUint32(boxes[0].length + 12), 8); // data offset (moof length + mdat length + mdat)
 
     return FragmentedISOBMFFBuilder.getBoxContents(boxes);
-  }
-
-  static getBoxContents(boxes) {
-    return Uint8Array.from(boxes.flatMap((box) => [...box.contents]));
   }
 }
