+++ src/demo/src/FragmentedMPEG/mpeg/Frame.js
@@ -1,0 +1,30 @@
+/* Copyright 2020 Ethan Halsall
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <https://www.gnu.org/licenses/>
+*/
+
+export default class Frame {
+  constructor(header, data) {
+    this._header = header;
+    this._data = data;
+  }
+
+  get header() {
+    return this._header;
+  }
+
+  get data() {
+    return this._data;
+  }
+}
