--- src/demo/src/isobmff/mpeg/Frame.js
@@ -1,14 +1,0 @@
-export default class Frame {
-  constructor(header, data) {
-    this._header = header;
-    this._data = data;
-  }
-
-  get header() {
-    return this._header;
-  }
-
-  get data() {
-    return this._data;
-  }
-}
